+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+##
+
+## Fixed
+
+- Request has scheme of `:http` when transmitted over `tcp` connection.
+
 ## [0.15.7](https://github.com/CrowdHailer/Ace/tree/0.15.7) - 2017-12-28
 
 ## Added